organize asset path
